els-ASAP auto commit file/tag gnutls-2.12.23-21.el6.src.rpm

+++ src/tests/utils.c
@@ -1,27 +1,28 @@
 /*
- * Copyright (C) 2004, 2005, 2006, 2007, 2008 Free Software Foundation
+ * Copyright (C) 2004, 2005, 2006, 2007, 2008, 2010 Free Software
+ * Foundation, Inc.
  *
  * Author: Simon Josefsson
  *
- * This file is part of GNUTLS.
+ * This file is part of GnuTLS.
  *
- * GNUTLS is free software; you can redistribute it and/or modify it
+ * GnuTLS is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 3 of the License, or
  * (at your option) any later version.
  *
- * GNUTLS is distributed in the hope that it will be useful, but
+ * GnuTLS is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with GNUTLS; if not, write to the Free Software Foundation,
+ * along with GnuTLS; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
  */
 
 #ifdef HAVE_CONFIG_H
-# include "config.h"
+#include "config.h"
 #endif
 
 #include <stdio.h>
@@ -68,18 +69,18 @@ escapeprint (const char *str, size_t len)
 {
   size_t i;
 
-  printf (" (length %d bytes):\n\t", len);
+  printf (" (length %d bytes):\n\t", (int) len);
   for (i = 0; i < len; i++)
     {
       if (((str[i] & 0xFF) >= 'A' && (str[i] & 0xFF) <= 'Z') ||
-	  ((str[i] & 0xFF) >= 'a' && (str[i] & 0xFF) <= 'z') ||
-	  ((str[i] & 0xFF) >= '0' && (str[i] & 0xFF) <= '9')
-	  || (str[i] & 0xFF) == ' ' || (str[i] & 0xFF) == '.')
-	printf ("%c", (str[i] & 0xFF));
+          ((str[i] & 0xFF) >= 'a' && (str[i] & 0xFF) <= 'z') ||
+          ((str[i] & 0xFF) >= '0' && (str[i] & 0xFF) <= '9')
+          || (str[i] & 0xFF) == ' ' || (str[i] & 0xFF) == '.')
+        printf ("%c", (str[i] & 0xFF));
       else
-	printf ("\\x%02X", (str[i] & 0xFF));
+        printf ("\\x%02X", (str[i] & 0xFF));
       if ((i + 1) % 16 == 0 && (i + 1) < len)
-	printf ("'\n\t'");
+        printf ("'\n\t'");
     }
   printf ("\n");
 }
@@ -94,9 +95,9 @@ hexprint (const char *str, size_t len)
     {
       printf ("%02x ", (str[i] & 0xFF));
       if ((i + 1) % 8 == 0)
-	printf (" ");
+        printf (" ");
       if ((i + 1) % 16 == 0 && i + 1 < len)
-	printf ("\n\t;; ");
+        printf ("\n\t;; ");
     }
   printf ("\n");
 }
@@ -110,17 +111,17 @@ binprint (const char *str, size_t len)
   for (i = 0; i < len; i++)
     {
       printf ("%d%d%d%d%d%d%d%d ",
-	      (str[i] & 0xFF) & 0x80 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x40 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x20 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x10 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x08 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x04 ? 1 : 0,
-	      (str[i] & 0xFF) & 0x02 ? 1 : 0, (str[i] & 0xFF) & 0x01 ? 1 : 0);
+              (str[i] & 0xFF) & 0x80 ? 1 : 0,
+              (str[i] & 0xFF) & 0x40 ? 1 : 0,
+              (str[i] & 0xFF) & 0x20 ? 1 : 0,
+              (str[i] & 0xFF) & 0x10 ? 1 : 0,
+              (str[i] & 0xFF) & 0x08 ? 1 : 0,
+              (str[i] & 0xFF) & 0x04 ? 1 : 0,
+              (str[i] & 0xFF) & 0x02 ? 1 : 0, (str[i] & 0xFF) & 0x01 ? 1 : 0);
       if ((i + 1) % 3 == 0)
-	printf (" ");
+        printf (" ");
       if ((i + 1) % 6 == 0 && i + 1 < len)
-	printf ("\n\t;; ");
+        printf ("\n\t;; ");
     }
   printf ("\n");
 }
@@ -130,18 +131,18 @@ main (int argc, char *argv[])
 {
   do
     if (strcmp (argv[argc - 1], "-v") == 0 ||
-	strcmp (argv[argc - 1], "--verbose") == 0)
+        strcmp (argv[argc - 1], "--verbose") == 0)
       debug = 1;
     else if (strcmp (argv[argc - 1], "-b") == 0 ||
-	     strcmp (argv[argc - 1], "--break-on-error") == 0)
+             strcmp (argv[argc - 1], "--break-on-error") == 0)
       break_on_error = 1;
     else if (strcmp (argv[argc - 1], "-h") == 0 ||
-	     strcmp (argv[argc - 1], "-?") == 0 ||
-	     strcmp (argv[argc - 1], "--help") == 0)
+             strcmp (argv[argc - 1], "-?") == 0 ||
+             strcmp (argv[argc - 1], "--help") == 0)
       {
-	printf ("Usage: %s [-vbh?] [--verbose] [--break-on-error] [--help]\n",
-		argv[0]);
-	return 1;
+        printf ("Usage: %s [-vbh?] [--verbose] [--break-on-error] [--help]\n",
+                argv[0]);
+        return 1;
       }
   while (argc-- > 1);
